Split tests for model and add test to validate ecosystem case insensitivity
